+++ Makefile
@@ -1,29 +1,27 @@
 CC = gcc
-CCFLAGS = -O2 -Wall -march=native -D_XOPEN_SOURCE -D_GNU_SOURCE -Wno-unused-variable
-
+CCFLAGS = -c -O2 -Wall -march=native -D_XOPEN_SOURCE -D_GNU_SOURCE -Wno-unused-variable
 LDFLAGS = $(shell pkg-config --libs ncurses)
 
-APP = todoApp
+BIN = bin/
+TMP = $(BIN)build/
+SRC = src/
+
+APP        = $(BIN)todoApp
+TODO_O     = $(TMP)todo.o
+GRAPHICS_O = $(TMP)graphics.o
+.PHONY:      clean run app
 
 
-OBJECTS = color.o resize_term.o todo.o
-.PHONY: clean run app
+$(APP): $(TODO_O) $(GRAPHICS_O)
+	$(CC) $(TODO_O) $(GRAPHICS_O) $(LDFLAGS) -o $(APP)
 
 run: $(APP)
 	./$(APP)
 
-$(APP): $(OBJECTS)
-	$(CC) $^ $(LDFLAGS) -o $(APP)
-
-
-.c.o:
-	$(CC) $(CCFLAGS) -c $<
-
-#dependencies
-todo.o: todo.c todo.h color.h
-color.o: color.c todo.h
-resize_term.o: resize_term.c color.h todo.h
-
+$(TODO_O): $(SRC)todo.c $(SRC)todo.h
+	$(CC) $(CCFLAGS) $(SRC)todo.c -o $(TODO_O)
+$(GRAPHICS_O): $(SRC)graphics.c
+	$(CC) $(CCFLAGS) $(SRC)graphics.c -o $(GRAPHICS_O)
 
 clean:
-	rm -rf $(OBJECTS) $(APP)
+	rm -rf $(TMP)* $(APP)
