--- color.h
@@ -1,10 +1,0 @@
-#ifndef COLOR_H
-#define COLOR_H
-
-int getTextColorNum(const char *txtColor);
-int getBkgdColorNum(const char *bkgdColor);
-
-void ncursesColorInit();
-
-
-#endif /*COLOR_H*/
