--- resize_term.h
@@ -1,7 +1,0 @@
-#ifndef RESIZE_TERM_H
-#define RESIZE_TERM_H
-
-
-int setSigwinchHandler();
-
-#endif /*RESIZE_TERM_H*/
